Move is_compile_time_integer to sdfg_utils module
