+++ src/gt4py/next/program_processors/runners/dace/sdfg_callable.py
@@ -15,7 +15,10 @@ import dace
 
 from gt4py._core import definitions as core_defs
 from gt4py.next import common as gtx_common, field_utils
-from gt4py.next.program_processors.runners.dace import sdfg_args as gtx_dace_args
+from gt4py.next.program_processors.runners.dace import (
+    sdfg_args as gtx_dace_args,
+    sdfg_utils as gtx_dace_utils,
+)
 
 
 def get_field_domain_symbols(name: str, domain: gtx_common.Domain) -> dict[str, int]:
@@ -34,7 +37,7 @@ def get_array_shape_symbols(
 ) -> dict[str, int]:
     array_symbols = {}
     for array_size, size in zip(array_desc.shape, ndarray.shape, strict=True):
-        if gtx_dace_args.is_compile_time_integer(array_size):
+        if gtx_dace_utils.is_compile_time_integer(array_size):
             if int(array_size) != size:
                 raise RuntimeError(f"Array shape mismatch: expected {array_size}, got {size}.")
         else:
@@ -50,7 +53,7 @@ def get_array_stride_symbols(
     for array_stride, value in zip(array_desc.strides, ndarray.strides, strict=True):
         assert divmod(value, ndarray.itemsize)[1] == 0
         stride = value // ndarray.itemsize
-        if gtx_dace_args.is_compile_time_integer(array_stride):
+        if gtx_dace_utils.is_compile_time_integer(array_stride):
             if int(array_stride) != stride:
                 raise RuntimeError(f"Array stride mismatch: expected {array_stride}, got {stride}.")
         else:
