+++ src/gt4py/next/program_processors/runners/dace/transformations/loop_blocking.py
@@ -33,7 +33,7 @@ from gt4py.next import common as gtx_common
 from gt4py.next.program_processors.runners.dace import (
     library_nodes as gtx_lib,
     lowering as gtx_dace_lowering,
-    sdfg_args as gtx_dace_args,
+    sdfg_utils as gtx_dace_utils,
 )
 
 
@@ -783,7 +783,7 @@ class LoopBlocking(dace_transformation.SingleStateTransformation):
 
         promote_subset: dace_subsets.Subset = edge.data.subset  # Works because of canonicalization.
         buffer_shape = tuple(dace_symb.overapproximate(promote_subset.size()))
-        if not all(gtx_dace_args.is_compile_time_integer(x) for x in buffer_shape):
+        if not all(gtx_dace_utils.is_compile_time_integer(x) for x in buffer_shape):
             return False
         for buffer_dim in buffer_shape:
             # We don't want to promote memlets that are too big because then the register or stack usage would increase a lot and we will get worse performance.
