+++ src/gt4py/next/program_processors/runners/dace/transformations/map_promoter.py
@@ -19,7 +19,7 @@ from dace import (
 from dace.sdfg import nodes as dace_nodes
 
 from gt4py.next.program_processors.runners.dace import (
-    sdfg_args as gtx_dace_args,
+    sdfg_utils as gtx_dace_utils,
     transformations as gtx_transformations,
 )
 
@@ -273,7 +273,7 @@ class MapPromoter(dace_transformation.SingleStateTransformation):
         #  To prevent that we require that the number of iterations the second Map
         #  performs, at compile time, is larger than zero.
         second_map_iterations: Any = second_map_entry.map.range.num_elements()
-        if gtx_dace_args.is_compile_time_integer(second_map_iterations):
+        if gtx_dace_utils.is_compile_time_integer(second_map_iterations):
             second_map_iterations = int(str(second_map_iterations))
             if second_map_iterations <= 0:
                 return False
